pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code_oracle"
version = "0.1.0"
description = "Code location tool using Gemini 2.5 Flash"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "JP", email = "user@example.com"},
]

dependencies = [
    "langchain-core",
    "langchain",
    "pydantic",
    "google-generativeai",
    "anthropic>=0.49.0",
    "langchain-anthropic>=0.3.12",
    "langgraph-prebuilt>=0.1.8",
    "python-dotenv",
    "fastmcp>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pytest",
    "pytest-asyncio",
]

[tool.setuptools]
packages = ["app", "config"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true
disable_error_code = ["attr-defined", "name-defined", "misc"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto" 
